% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data set}
\format{
## `fire`
A data frame with 15 rows and 2 variables:
\describe{
\item{Distance}{distance from fire station in miles}
\item{DAMAGE}{X 1000 dollars in damage}
\item{year}{Year}
...
}
}
\source{
\url{http://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendhenhall-Sincish/p/book/9781498728850}
}
\usage{
fire
}
\description{
A subset of data that shows the distance and damage of a fire
Report ...
}
\keyword{datasets}
