% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of trials}

\item{p}{Probability}
}
\value{
A table with the proportions
}
\description{
Binomial
}
\examples{
mybin(1000, 10, 0.7)

}
