% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Condidence Interval Estimator with Histogram}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iteration}

\item{x}{A numeric vecotr of data values}

\item{fun}{A funcation or the name of a function to compute the statistic of interest}

\item{alpha}{Significance level for the confidence interval}

\item{cx}{Text size for annotation on the plot}

\item{...}{}
}
\value{
A list containing ci, fun, x, xstats
}
\description{
Bootstrap Condidence Interval Estimator with Histogram
}
\examples{
set.seed(123)
x <- rnorm(20, mean = 5, sd = 2)
myboot2(x = x, fun = "mean")
}
