% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Generate and Plot a Binomial Sampling Distribution}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{Sample size for each iteration}

\item{iter}{Number of iterations}

\item{p}{Probability of success in the binomial distribution}

\item{...}{}
}
\value{
A vector of sample means from the binomial distribution
}
\description{
Generate and Plot a Binomial Sampling Distribution
}
\examples{
mycltb(n=10, iter=10000, p=0.3)
}
