% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Maximum Likelihood estimation for normal distribution}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{Sample data}

\item{mu}{Mean values}

\item{sig}{Standard deviation}

\item{...}{}
}
\value{
Maximum likelihood value
}
\description{
Maximum Likelihood estimation for normal distribution
}
\examples{
mymlnorm(x=c(0,12,13,15,12,11,10),mu=seq(5,15,length=1000),sig=seq(0.1,6,length=1000),lwd=2,labcex=1)

}
