% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot Normal Distribution Curve with Shaded Probability Area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{a}{The x-value where the area is shaded}
}
\value{
A graph with the curve of the normal distribution and the shaded area
}
\description{
Plot Normal Distribution Curve with Shaded Probability Area
}
\examples{
myncurve(mu = 10, sigma = 2, a = 12)
}
