% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R, R/project1.R
\name{ntickets}
\alias{ntickets}
\title{Calcuate the number of tickets to sell using binomial and normal disribution}
\usage{
ntickets(N, gamma, p)

ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of available seats}

\item{gamma}{Maximum acceptable probabilily of overbooking}

\item{p}{Probability that a single passenger shows up for the flight}
}
\value{
A list of the inputs and the plots of the discrete and normal distributions

A list of the inputs and the plots of the discrete and normal distributions
}
\description{
Calcuate the number of tickets to sell using binomial and normal disribution

Calcuate the number of tickets to sell using binomial and normal disribution
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)
ntickets(N=400,gamma = 0.02, p = 0.95)
}
